Micronaut: loadtest works.

+++ micronaut-petclinic-customers-service/src/main/java/org/micronautframework/samples/petclinic/customers/model/PetRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 
 @Repository
-public interface PetRepository extends CrudRepository<Pet, Long> {
+public interface PetRepository extends CrudRepository<Pet, Integer> {
     @Query("SELECT ptype FROM PetType ptype ORDER BY ptype.name")
     List<PetType> findPetTypes();
