+++ micronaut-petclinic-visits-service/src/main/java/org/micronautframework/samples/petclinic/visits/model/VisitRepository.java
@@ -8,7 +8,7 @@ import java.util.Collection;
 import java.util.List;
 
 @Repository
-public interface VisitRepository extends CrudRepository<Visit, Long> {
+public interface VisitRepository extends CrudRepository<Visit, Integer> {
     @Executable
     List<Visit> findByPetId(int petId);
